pyproject.toml
[project]
name = "activation-lab"
version = "0.1.0"
description = "Capture per-step LLM activations for research and interpretability"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3",
    "transformers>=4.45",
    "accelerate>=0.30",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "typer>=0.12",
    "rich>=13.7",
    "matplotlib>=3.8",
    "streamlit>=1.35",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]
notebook = [
    "jupyter>=1.0",
    "ipykernel>=6.29",
]

[project.scripts]
activation-lab = "activation_lab.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/activation_lab"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]
